refactor: ~UseCaseService (#6)
