+++ woorizip-core/src/main/kotlin/core/CoreModuleBinding.kt
@@ -1,22 +1,20 @@
 package core
 
-import core.meeting.service.CheckMeetingService
-import core.meeting.service.CheckMeetingServiceImpl
-import core.meeting.service.CommandMeetingExternalService
-import core.meeting.service.CommandMeetingExternalServiceImpl
 import core.meeting.service.CommandMeetingService
 import core.meeting.service.CommandMeetingServiceImpl
 import core.meeting.service.MeetingService
+import core.meeting.service.MeetingUseCaseService
+import core.meeting.service.MeetingUseCaseServiceImpl
 import core.meeting.service.QueryMeetingService
 import core.meeting.service.QueryMeetingServiceImpl
 import core.meeting.usecase.CreateMeeting
 import core.member.service.CheckMemberService
 import core.member.service.CheckMemberServiceImpl
-import core.member.service.CommandMemberExternalService
-import core.member.service.CommandMemberExternalServiceImpl
 import core.member.service.CommandMemberService
 import core.member.service.CommandMemberServiceImpl
 import core.member.service.MemberService
+import core.member.service.MemberUseCaseService
+import core.member.service.MemberUseCaseServiceImpl
 import core.member.service.QueryMemberService
 import core.member.service.QueryMemberServiceImpl
 import core.member.usecase.SignIn
@@ -34,17 +32,16 @@ val coreModule: Module
 internal val serviceModule = module {
     // member
     singleOf(::MemberService)
+    singleOf(::MemberUseCaseServiceImpl) bind MemberUseCaseService::class
     singleOf(::QueryMemberServiceImpl) bind QueryMemberService::class
     singleOf(::CommandMemberServiceImpl) bind CommandMemberService::class
     singleOf(::CheckMemberServiceImpl) bind CheckMemberService::class
-    singleOf(::CommandMemberExternalServiceImpl) bind CommandMemberExternalService::class
 
     // meeting
     singleOf(::MeetingService)
+    singleOf(::MeetingUseCaseServiceImpl) bind MeetingUseCaseService::class
     singleOf(::QueryMeetingServiceImpl) bind QueryMeetingService::class
     singleOf(::CommandMeetingServiceImpl) bind CommandMeetingService::class
-    singleOf(::CommandMeetingExternalServiceImpl) bind CommandMeetingExternalService::class
-    singleOf(::CheckMeetingServiceImpl) bind CheckMeetingService::class
 }
 
 internal val useCaseModule = module {
